web-ui/requirements.txt
streamlit
requests
pandas
pillow
nest-asyncio
pyyaml
boto3
